Responded to code reviewing questions
