+++ src/elm/State.elm
@@ -44,6 +44,9 @@ update msg model =
             { model | view = getViewFromRoute location.hash }
                 ! (scrollToTop :: viewToCmds model.view)
 
+        ChangeView view ->
+            { model | view = view } ! []
+
         RecieveHotspotCoords (Ok coords) ->
             { model | hotspots = coords } ! []
