+++ src/js/elm-init.js
@@ -1,0 +1,4 @@
+import Elm from '../elm/Main.elm';
+const app = Elm.Main.fullscreen();
+
+export default app;
